# Stop persisting current_loc (+2% parse throughput)