--- a/src/generic/stage2/structural_parser.h
+++ b/src/generic/stage2/structural_parser.h
@@ -6,6 +6,8 @@
 namespace stage2 {
 namespace { // Make everything here private
 
+#include "generic/stage2/tape_writer.h"
+
 #ifdef SIMDJSON_USE_COMPUTED_GOTO
 #define INIT_ADDRESSES() { &&array_begin, &&array_continue, &&error, &&finish, &&object_begin, &&object_continue }
 #define GOTO(address) { goto *(address); }
@@ -46,45 +48,25 @@ struct unified_machine_addresses {
 #undef FAIL_IF
 #define FAIL_IF(EXPR) { if (EXPR) { return addresses.error; } }
 
-struct number_writer {
-  dom_parser_implementation &parser;
-  
-  really_inline void write_s64(int64_t value) noexcept {
-    append_tape(0, internal::tape_type::INT64);
-    std::memcpy(&parser.doc->tape[parser.current_loc], &value, sizeof(value));
-    ++parser.current_loc;
-  }
-  really_inline void write_u64(uint64_t value) noexcept {
-    append_tape(0, internal::tape_type::UINT64);
-    parser.doc->tape[parser.current_loc++] = value;
-  }
-  really_inline void write_double(double value) noexcept {
-    append_tape(0, internal::tape_type::DOUBLE);
-    static_assert(sizeof(value) == sizeof(parser.doc->tape[parser.current_loc]), "mismatch size");
-    memcpy(&parser.doc->tape[parser.current_loc++], &value, sizeof(double));
-    // doc->tape[doc->current_loc++] = *((uint64_t *)&d);
-  }
-  really_inline void append_tape(uint64_t val, internal::tape_type t) noexcept {
-    parser.doc->tape[parser.current_loc++] = val | ((uint64_t(char(t))) << 56);
-  }
-}; // struct number_writer
-
 struct structural_parser : structural_iterator {
+  /** Lets you append to the tape */
+  tape_writer tape;
   /** Next write location in the string buf for stage 2 parsing */
-  uint8_t *current_string_buf_loc{};
+  uint8_t *current_string_buf_loc;
   /** Current depth (nested objects and arrays) */
-  uint32_t depth;
+  uint32_t depth{0};
 
   // For non-streaming, to pass an explicit 0 as next_structural, which enables optimizations
   really_inline structural_parser(dom_parser_implementation &_parser, uint32_t start_structural_index)
     : structural_iterator(_parser, start_structural_index),
-      depth{0} {
+      tape{parser.doc->tape.get()},
+      current_string_buf_loc{parser.doc->string_buf.get()} {
   }
 
   WARN_UNUSED really_inline bool start_scope(ret_address_t continue_state) {
-    parser.containing_scope[depth].tape_index = parser.current_loc;
+    parser.containing_scope[depth].tape_index = next_tape_index();
     parser.containing_scope[depth].count = 0;
-    parser.current_loc++; // We don't actually *write* the start element until the end.
+    tape.skip(); // We don't actually *write* the start element until the end.
     parser.ret_address[depth] = continue_state;
     depth++;
     bool exceeded_max_depth = depth >= parser.max_depth();
@@ -112,14 +94,18 @@ struct structural_parser : structural_iterator {
     depth--;
     // write our doc->tape location to the header scope
     // The root scope gets written *at* the previous location.
-    append_tape(parser.containing_scope[depth].tape_index, end);
+    tape.append(parser.containing_scope[depth].tape_index, end);
     // count can overflow if it exceeds 24 bits... so we saturate
     // the convention being that a cnt of 0xffffff or more is undetermined in value (>=  0xffffff).
     const uint32_t start_tape_index = parser.containing_scope[depth].tape_index;
     const uint32_t count = parser.containing_scope[depth].count;
     const uint32_t cntsat = count > 0xFFFFFF ? 0xFFFFFF : count;
     // This is a load and an OR. It would be possible to just write once at doc->tape[d.tape_index]
-    write_tape(start_tape_index, parser.current_loc | (uint64_t(cntsat) << 32), start);
+    tape_writer::write(parser.doc->tape[start_tape_index], next_tape_index() | (uint64_t(cntsat) << 32), start);
+  }
+
+  really_inline uint32_t next_tape_index() {
+    return uint32_t(tape.next_tape_loc - parser.doc->tape.get());
   }
 
   really_inline void end_object() {
@@ -135,14 +121,6 @@ struct structural_parser : structural_iterator {
     end_scope(internal::tape_type::ROOT, internal::tape_type::ROOT);
   }
 
-  really_inline void append_tape(uint64_t val, internal::tape_type t) noexcept {
-    parser.doc->tape[parser.current_loc++] = val | ((uint64_t(char(t))) << 56);
-  }
-
-  really_inline void write_tape(uint32_t loc, uint64_t val, internal::tape_type t) noexcept {
-    parser.doc->tape[loc] = val | ((uint64_t(char(t))) << 56);
-  }
-
   // increment_count increments the count of keys in an object or values in an array.
   // Note that if you are at the level of the values or elements, the count
   // must be increment in the preceding depth (depth-1) where the array or
@@ -153,7 +131,7 @@ struct structural_parser : structural_iterator {
 
   really_inline uint8_t *on_start_string() noexcept {
     // we advance the point, accounting for the fact that we have a NULL termination
-    append_tape(current_string_buf_loc - parser.doc->string_buf.get(), internal::tape_type::STRING);
+    tape.append(current_string_buf_loc - parser.doc->string_buf.get(), internal::tape_type::STRING);
     return current_string_buf_loc + sizeof(uint32_t);
   }
 
@@ -183,8 +161,7 @@ struct structural_parser : structural_iterator {
 
   WARN_UNUSED really_inline bool parse_number(const uint8_t *src, bool found_minus) {
     log_value("number");
-    number_writer writer{parser};
-    bool succeeded = numberparsing::parse_number(src, found_minus, writer);
+    bool succeeded = numberparsing::parse_number(src, found_minus, tape);
     if (!succeeded) { log_error("Invalid number"); }
     return !succeeded;
   }
@@ -200,17 +177,17 @@ struct structural_parser : structural_iterator {
     case 't':
       log_value("true");
       FAIL_IF( !atomparsing::is_valid_true_atom(current()) );
-      append_tape(0, internal::tape_type::TRUE_VALUE);
+      tape.append(0, internal::tape_type::TRUE_VALUE);
       return continue_state;
     case 'f':
       log_value("false");
       FAIL_IF( !atomparsing::is_valid_false_atom(current()) );
-      append_tape(0, internal::tape_type::FALSE_VALUE);
+      tape.append(0, internal::tape_type::FALSE_VALUE);
       return continue_state;
     case 'n':
       log_value("null");
       FAIL_IF( !atomparsing::is_valid_null_atom(current()) );
-      append_tape(0, internal::tape_type::NULL_VALUE);
+      tape.append(0, internal::tape_type::NULL_VALUE);
       return continue_state;
     case '0': case '1': case '2': case '3': case '4':
     case '5': case '6': case '7': case '8': case '9':
@@ -285,8 +262,6 @@ struct structural_parser : structural_iterator {
 
   really_inline void init() {
     log_start();
-    current_string_buf_loc = parser.doc->string_buf.get();
-    parser.current_loc = 0;
     parser.error = UNINITIALIZED;
   }
 
@@ -362,17 +337,17 @@ WARN_UNUSED static error_code parse_structurals(dom_parser_implementation &dom_p
   case 't':
     parser.log_value("true");
     FAIL_IF( !atomparsing::is_valid_true_atom(parser.current(), parser.remaining_len()) );
-    parser.append_tape(0, internal::tape_type::TRUE_VALUE);
+    parser.tape.append(0, internal::tape_type::TRUE_VALUE);
     goto finish;
   case 'f':
     parser.log_value("false");
     FAIL_IF( !atomparsing::is_valid_false_atom(parser.current(), parser.remaining_len()) );
-    parser.append_tape(0, internal::tape_type::FALSE_VALUE);
+    parser.tape.append(0, internal::tape_type::FALSE_VALUE);
     goto finish;
   case 'n':
     parser.log_value("null");
     FAIL_IF( !atomparsing::is_valid_null_atom(parser.current(), parser.remaining_len()) );
-    parser.append_tape(0, internal::tape_type::NULL_VALUE);
+    parser.tape.append(0, internal::tape_type::NULL_VALUE);
     goto finish;
   case '0': case '1': case '2': case '3': case '4':
   case '5': case '6': case '7': case '8': case '9':