This change makes it more likely the compiler will enregister the next tape write location (a very hot variable) by making it temporary to stage 2 (i.e. in structural_parser) instead of persisting it in dom_parser_implementation.

It looks like an overall average throughput improvement close to 2%. Instruction count is down most places. Throughput improvement happens on all but 4 files on skylake:

| File | Old Cycles | New Cycles | +% | Old Ins. | New Ins. | +% |
| --- | --: | --: | --: | --: | --: | --: |
| mesh | 275 | 265 | 4% | 877 | 856 | 2% |
| marine_ik | 273 | 265 | 3% | 821 | 806 | 2% |
| twitter_api_compact_response | 124 | 121 | 2% | 389 | 379 | 3% |
| apache_builds | 94 | 92 | 2% | 319 | 315 | 1% |
| numbers | 231 | 227 | 2% | 725 | 717 | 1% |
| random | 146 | 143 | 2% | 496 | 492 | 1% |
| twitter_api_response | 100 | 98 | 2% | 318 | 310 | 2% |
| github_events | 86 | 85 | 1% | 278 | 273 | 2% |
| instruments | 101 | 99 | 1% | 352 | 352 | 0% |
| update-center | 117 | 116 | 1% | 352 | 346 | 2% |
| twitter | 95 | 94 | 1% | 307 | 299 | 2% |
| google_maps_api_response | 111 | 111 | 1% | 365 | 367 | -1% |
| gsoc-2018 | 70 | 70 | 0% | 178 | 176 | 1% |
| twitter_timeline | 134 | 134 | 0% | 411 | 400 | 3% |
| repeat | 113 | 113 | 0% | 364 | 366 | -1% |
| tree-pretty | 102 | 102 | 0% | 346 | 343 | 1% |
| twitterescaped | 195 | 196 | 0% | 560 | 563 | 0% |
| mesh.pretty | 166 | 167 | -1% | 533 | 524 | 2% |
| google_maps_api_compact_response | 206 | 208 | -1% | 682 | 687 | -1% |
| citm_catalog | 84 | 85 | -1% | 299 | 304 | -2% |
| canada | 263 | 284 | -8% | 916 | 906 | 1% |

### PR Notes:

* This change doesn't have its awesome effect on master (it makes it worse). It doesn't get really good until #918--presumably because removing several variables from structural_iterator reduced reduced the number of decisions the compiler had to make about registers.
* number_writer: After moving current_loc to the structural_parser, the number_writer couldn't get to it anymore.
  - I couldn't change the `dom_parser_implementation &parser` to `structural_parser &parser`, because number_writer gets passed to parse_large_integer and slow_parse_float, and that would deenregister the whole thing.
  - So I changed current_loc from uint32_t to `uint64_t *` and had number_writer store *that*. That single pointer is enough to write values to the right place *and* update the pointer.
  - Unfortunately, that wasn't enough: any attempt to pass `next_tape_loc` by reference or pointer caused it to de-enregister (reasonably enough). So I pass it by value, let it write the number out, and then bump next_tape_loc after calling either slow_parse_float or parse_large_integer.
* I also changed number_writer to tape_writer and had structural_parser use it everywhere, so the write logic is largely localized.